move method out of class
