+++ openhands/utils/conversation_summary.py
@@ -4,7 +4,12 @@ from typing import Optional
 
 from openhands.core.config import LLMConfig
 from openhands.core.logger import openhands_logger as logger
+from openhands.events.action.message import MessageAction
+from openhands.events.event import EventSource
+from openhands.events.stream import EventStream
 from openhands.llm.llm import LLM
+from openhands.storage.data_models.settings import Settings
+from openhands.storage.files import FileStore
 
 
 async def generate_conversation_title(
@@ -68,3 +73,68 @@ def get_default_conversation_title(conversation_id: str) -> str:
         A default title string
     """
     return f'Conversation {conversation_id[:5]}'
+
+
+async def auto_generate_title(
+    conversation_id: str, user_id: str | None, file_store: FileStore, settings: Settings
+) -> str:
+    """
+    Auto-generate a title for a conversation based on the first user message.
+    Uses LLM-based title generation if available, otherwise falls back to a simple truncation.
+
+    Args:
+        conversation_id: The ID of the conversation
+        user_id: The ID of the user
+
+    Returns:
+        A generated title string
+    """
+    logger.info(f'Auto-generating title for conversation {conversation_id}')
+
+    try:
+        # Create an event stream for the conversation
+        event_stream = EventStream(conversation_id, file_store, user_id)
+
+        # Find the first user message
+        first_user_message = None
+        for event in event_stream.get_events():
+            if (
+                event.source == EventSource.USER
+                and isinstance(event, MessageAction)
+                and event.content
+                and event.content.strip()
+            ):
+                first_user_message = event.content
+                break
+
+        if first_user_message:
+            # Get LLM config from user settings
+            try:
+                if settings and settings.llm_model:
+                    # Create LLM config from settings
+                    llm_config = LLMConfig(
+                        model=settings.llm_model,
+                        api_key=settings.llm_api_key,
+                        base_url=settings.llm_base_url,
+                    )
+
+                    # Try to generate title using LLM
+                    llm_title = await generate_conversation_title(
+                        first_user_message, llm_config
+                    )
+                    if llm_title:
+                        logger.info(f'Generated title using LLM: {llm_title}')
+                        return llm_title
+            except Exception as e:
+                logger.error(f'Error using LLM for title generation: {e}')
+
+            # Fall back to simple truncation if LLM generation fails or is unavailable
+            first_user_message = first_user_message.strip()
+            title = first_user_message[:30]
+            if len(first_user_message) > 30:
+                title += '...'
+            logger.info(f'Generated title using truncation: {title}')
+            return title
+    except Exception as e:
+        logger.error(f'Error generating title: {str(e)}')
+    return ''
